Refactoring of the scaler
