+++ CODEOWNERS
@@ -11,5 +11,6 @@
 **/llpr                 @ppegolo @SanggyuChong
 **/mace                 @pfebrer @jwa7
 **/pet                  @abmazitov @sofiia-chorna
+**/scaler               @pfebrer
 **/soap_bpnn            @ppegolo
 **/space                @frostedoyster
