+++ docs/src/dev-docs/changelog.rst
@@ -56,15 +56,13 @@ Added
   directory can be read back directly as a ``systems: read_from:`` dataset, which is
   useful for very large evaluation runs.
 - MACE architecture now supports multi-headed MACE models through the ``mace_head_name`` hyperparameter.
-- ``composition`` is now a standalone architecture: it can be trained, exported, and run
-  for inference on its own (``architecture: {name: composition}``), in addition to being
-  used as an additive baseline inside the other architectures. Standalone training
-  supports distributed environments (detected automatically from the SLURM
-  environment), fits on the requested device, and loads data in parallel
-  (``num_workers``).
-- The ``atomic_baseline`` hyperparameter now also accepts a path to a pretrained
-  composition checkpoint, which is loaded and reused as the additive baseline instead of
-  being refitted from the training data.
+- Moved composition model and scaler to standalone architectures.
+  They can be trained, exported, and run for inference on their own
+  (``architecture: {name: composition | scaler}``), in addition to being
+  used inside the other architectures.
+- The ``atomic_baseline`` and ``fixed_scaling_weights`` hyperparameters now also accept
+  paths to pretrained composition/scaler checkpoints, which are loaded and reused instead
+  of being refitted from the training data.
 - Rotational augmentation now supports atomic-basis targets and Cartesian targets of
   rank > 2.
 - Atom-pair (``sample_kind: atom_pair``) targets are now supported for ``scalar``,
@@ -100,6 +98,11 @@ Changed
 - ``RotationalAugmenter`` is now ``O3Augmenter``, with a ``group`` option selecting
   the transformations to sample (``"O3"`` or ``"inversions"``). The PhACE
   ``InversionAugmenter`` was removed in its favor.
+- The scaler model moved from ``metatrain.utils.scaler.Scaler`` to
+  ``metatrain.scaler.Scaler``.
+- ``Scaler``'s forward method does no longer apply scales. Instead it returns the scales for
+  each target. To apply scales use ``Scaler.apply_scales``. This is because the previous
+  behavior was not compatible with the forward method of ``ModelInterface``.
 
 Removed
 #######
