+++ src/metatrain/pet/model.py
@@ -17,6 +17,7 @@ from metatomic.torch import (
 )
 
 from metatrain.composition import CompositionModel
+from metatrain.scaler import Scaler
 from metatrain.utils.abc import ModelInterface
 from metatrain.utils.additive import ZBL
 from metatrain.utils.data import DatasetInfo, TargetInfo
@@ -31,7 +32,6 @@ from metatrain.utils.data.atomic_basis_helpers import (
 from metatrain.utils.dtype import dtype_to_str
 from metatrain.utils.long_range import DummyLongRangeFeaturizer, LongRangeFeaturizer
 from metatrain.utils.metadata import merge_metadata
-from metatrain.utils.scaler import Scaler
 from metatrain.utils.sum_over_atoms import sum_over_atoms
 
 from . import checkpoints
@@ -590,7 +590,7 @@ class PET(ModelInterface[ModelHypers]):
         with torch.profiler.record_function("PET::post-processing"):
             if not self.training:
                 # at evaluation, we also introduce the scaler and additive contributions
-                return_dict = self.scaler(
+                return_dict = self.scaler.apply_scales(
                     systems,
                     return_dict,
                     selected_atoms=selected_atoms,
