+++ src/metatrain/utils/scaler/__init__.py
@@ -1,2 +1,1 @@
 from .remove import get_remove_scale_transform, remove_scale  # noqa: F401
-from .scaler import FixedScalerWeights, Scaler  # noqa: F401
