+++ tox.ini
@@ -16,6 +16,7 @@ envlist =
     llpr-tests
     classifier-tests
     composition-tests
+    scaler-tests
     mace-tests
     space-tests
     dpa3-tests
@@ -44,7 +45,8 @@ lint_strict_folders =
     "{toxinidir}/src/metatrain/pet" \
     "{toxinidir}/src/metatrain/gap" \
     "{toxinidir}/src/metatrain/soap_bpnn" \
-    "{toxinidir}/src/metatrain/composition"
+    "{toxinidir}/src/metatrain/composition" \
+    "{toxinidir}/src/metatrain/scaler"
 exclude_tests_regex = "(^|/)[^/]*test[^/]*$"
 exclude_tests_glob = "**/*test*"
 
@@ -244,6 +246,16 @@ commands =
 
 [testenv:composition-tests]
 description = Run composition model tests with pytest and collect coverage
+passenv = *
+deps =
+    pytest
+    pytest-cov
+changedir = src/metatrain/composition/tests/
+commands =
+    pytest --cov=metatrain.composition --cov-append --cov-report=term {posargs}
+
+[testenv:scaler-tests]
+description = Run scaler tests with pytest and collect coverage
 passenv = *
 deps =
     pytest
